Allow the 'username=' kwarg to be turned off
User-written view functions may expect this kwarg, but many others do
not. This change allows using require_auth() as a wrapper for view_funcs
which do not accept arbitrary kwargs.

(For example, jsonrpc.backend.flask.api.as_view() can now be wrapped.)

I would actually prefer this to be disabled by default, as REMOTE_USER
seems to be much more standard, but that would result in an API break.